Extract remediation-specific utility functions from build_remediations.py into a separate module (remediations.py)
